+++ src/site/Router.php
@@ -17,7 +17,7 @@ class Router {
       '/{product:(?:hack|php)}/reference/{type:(?:class|function|interface|trait)}/'
         => APIListController::class,
       '/{product:(?:hack|php)}/reference/{type:(?:class|function|interface|trait)}/{name}/'
-        => APIGenericPageController::class,
+        => APIClassPageController::class,
       '/{product:(?:hack)}/reference/{type:(?:class|interface|trait)}/{class}/{method}/'
         => APIMethodPageController::class,
       '/{product:(?:hack|hhvm)}/{guide}/'
